+++ source/common/http/conn_manager_impl.cc
@@ -376,10 +376,12 @@ ConnectionManagerImpl::ActiveStream::~ActiveStream() {
 
   connection_manager_.stats_.named_.downstream_rq_active_.dec();
   for (const AccessLog::InstanceSharedPtr& access_log : connection_manager_.config_.accessLogs()) {
-    access_log->log(request_headers_.get(), response_headers_.get(), request_info_);
+    access_log->log(request_headers_.get(), response_headers_.get(), response_trailers_.get(),
+                    request_info_);
   }
   for (const auto& log_handler : access_log_handlers_) {
-    log_handler->log(request_headers_.get(), response_headers_.get(), request_info_);
+    log_handler->log(request_headers_.get(), response_headers_.get(), response_trailers_.get(),
+                     request_info_);
   }
 
   if (request_info_.healthCheck()) {