+++ test/integration/tcp_proxy_integration_test.cc
@@ -263,6 +263,34 @@ TEST_P(TcpProxyIntegrationTest, AccessLog) {
                                        ip_port_regex, ip_regex)));
 }
 
+// Test that the server shuts down without crashing when connections are open.
+TEST_P(TcpProxyIntegrationTest, ShutdownWithOpenConnections) {
+  config_helper_.addConfigModifier([&](envoy::config::bootstrap::v2::Bootstrap& bootstrap) -> void {
+    auto* static_resources = bootstrap.mutable_static_resources();
+    for (int i = 0; i < static_resources->clusters_size(); ++i) {
+      auto* cluster = static_resources->mutable_clusters(i);
+      cluster->set_close_connections_on_host_health_failure(true);
+    }
+  });
+  initialize();
+  IntegrationTcpClientPtr tcp_client = makeTcpConnection(lookupPort("tcp_proxy"));
+  tcp_client->write("hello");
+  FakeRawConnectionPtr fake_upstream_connection = fake_upstreams_[0]->waitForRawConnection();
+  fake_upstream_connection->waitForData(5);
+  fake_upstream_connection->write("world");
+  tcp_client->waitForData("world");
+  tcp_client->write("hello", false);
+  fake_upstream_connection->waitForData(10);
+  test_server_.reset();
+  fake_upstream_connection->waitForHalfClose();
+  fake_upstream_connection->close();
+  fake_upstream_connection->waitForDisconnect(true);
+  tcp_client->waitForHalfClose();
+  tcp_client->close();
+
+  // Success criteria is that no ASSERTs fire and there are no leaks.
+}
+
 void TcpProxySslIntegrationTest::initialize() {
   config_helper_.addSslConfig();
   TcpProxyIntegrationTest::initialize();